+++ src/main/java/io/dataspaceconnector/service/appstore/portainer/PortainerService.java
@@ -798,7 +798,6 @@ public class PortainerService {
             final List<AppEndpointImpl> appEndpoints
     ) throws IOException {
         final var templateObject = toJsonObject(appStoreTemplate);
-        final var image = templateObject.getString("image");
 
         //get all ports from the appTemplate (with label)
         var portLabelMap = new HashMap<String, String>();
@@ -818,12 +817,15 @@ public class PortainerService {
             }
         }
 
+        final var containerName = "app-" + System.currentTimeMillis();
+
         final var builder = getRequestBuilder();
         final var urlBuilder = new HttpUrl.Builder()
                 .scheme(portainerConfig.getScheme())
                 .host(portainerConfig.getHost())
                 .port(portainerConfig.getPort())
-                .addPathSegments(API_ENDPOINT + endpointId + "/docker/containers/create");
+                .addPathSegments(API_ENDPOINT + endpointId + "/docker/containers/create")
+                .addQueryParameter("name", containerName);
 
         final var url = urlBuilder.build();
         builder.addHeader("Authorization", "Bearer " + getJwtToken());
@@ -835,8 +837,7 @@ public class PortainerService {
                         templateObject,
                         new ArrayList<>(portLabelMap.keySet()),
                         volumes,
-                        image
-                );
+                        containerName);
 
         //add json payload to request
         builder.post(